src/automate/RAG/requirements.txt
langchain
langchain-google-genai
google-generativeai

langchain_community
faiss-cpu
langchain-text-splitters
pymupdf
